Add POMYU chara loader For PMS JSON SKIN
